refactor(desktop): extract bearer pairing readiness
